requirements.txt
pandas==3.0.2
numpy==2.4.4
pyarrow==23.0.1
gymnasium==1.2.3
yfinance==1.2.0
openai>=1.0.0
python-dotenv>=1.0.0